refactor copyright component
